requirements.txt
numpy
pandas
matplotlib
scipy
scikit-learn
customtkinter
tensorly
pymcr
umap-learn
openpyxl